factor out tsPlugins
